+++ photutils/detection/__init__.py
@@ -5,5 +5,4 @@ an astronomical image.
 """
 
 from .core import *
-from .deblend import *
 from .findstars import *
